requirements.txt
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pymysql==1.1.0
python-dotenv==1.0.1
joblib==1.3.2
matplotlib==3.8.2
pyarrow==15.0.0